Add missing VZEROUPPER to memset-avx2
VZEROUPPER between AVX instructions and SSE instructions
is necessary in order to avoid performance degradations.

This patricular patch ensures vzerouppper is executed
when memset count is >64 and <128 bytes.

The guidance from Intel is that code that uses AVX that
returns/jumps must use VZEROUPPER in order to avoid
performance degradations.